Refactor service and repository classes for improved structure and maintainability

+++ src/main/java/unlp/info/bd2/services/ToursService.java
@@ -43,7 +43,7 @@ public interface ToursService {
     List<User> getTop5UsersMorePurchases();
     long getCountOfPurchasesBetweenDates(Date start, Date end);
     List<Route> getRoutesWithStop(Stop stop);
-    Long getMaxStopOfRoutes();
+    Integer getMaxStopOfRoutes();
     List<Route> getRoutsNotSell();
     List<Route> getTop3RoutesWithMaxRating();
     Service getMostDemandedService();
